--- src/pages/HomePage.tsx
@@ -1,7 +1,0 @@
-export const HomePage = () => {
-  return (
-    <div className="container">
-      <h1 className="title">Home page</h1>
-    </div>
-  );
-};
